добавлена модель с заглушкой repoImpl и вывод всего на экран
