--- app/src/main/java/com/example/weather/AppState.kt
@@ -1,7 +1,0 @@
-package com.example.weather
-
-sealed class AppState {
-    data class Success(val weatherData: Any) : AppState()
-    data class Error(val error: Throwable) : AppState()
-    object Loading : AppState()
-}
