--- app/src/main/java/com/example/weather/MainActivity.kt
@@ -1,22 +1,0 @@
-package com.example.weather
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.weather.databinding.MainActivityBinding
-import com.example.weather.ui.main.MainFragment
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: MainActivityBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = MainActivityBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        if (savedInstanceState == null) {
-            supportFragmentManager.beginTransaction()
-                    .replace(R.id.container, MainFragment.newInstance())
-                    .commitNow()
-        }
-    }
-}
