+++ app/src/main/java/com/example/weather/model/City.kt
@@ -1,0 +1,7 @@
+package com.example.weather.model
+
+data class City(
+    val city: String,
+    val lat: Double,
+    val lon: Double
+)
