+++ app/src/main/java/com/example/weather/model/Repo.kt
@@ -1,0 +1,6 @@
+package com.example.weather.model
+
+interface Repo {
+    fun getWeatherFromServer(): Weather
+    fun getWeatherFromLocalStorage(): Weather
+}
