+++ app/src/main/java/com/example/weather/model/RepoImpl.kt
@@ -1,0 +1,11 @@
+package com.example.weather.model
+
+class RepoImpl : Repo {
+    override fun getWeatherFromServer(): Weather {
+        return Weather()
+    }
+
+    override fun getWeatherFromLocalStorage(): Weather {
+        return Weather()
+    }
+}
