+++ app/src/main/java/com/example/weather/ui/viewmodel/AppState.kt
@@ -1,0 +1,9 @@
+package com.example.weather.ui.viewmodel
+
+import com.example.weather.model.Weather
+
+sealed class AppState {
+    data class Success(val weatherData: Weather) : AppState()
+    data class Error(val error: Throwable) : AppState()
+    object Loading : AppState()
+}
