+++ app/src/main/java/com/example/weather/ui/viewmodel/MainViewModel.kt
@@ -1,0 +1,25 @@
+package com.example.weather.ui.viewmodel
+
+import android.os.SystemClock.sleep
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.example.weather.model.Repo
+import com.example.weather.model.RepoImpl
+
+class MainViewModel(
+    private val liveDataToObserve: MutableLiveData<AppState> = MutableLiveData(),
+    private val repoImpl: Repo = RepoImpl()
+) :
+    ViewModel() {
+    fun getLiveData() = liveDataToObserve
+
+    fun getWeather() = getDataFromLocalSource()
+
+    private fun getDataFromLocalSource() {
+        liveDataToObserve.value = AppState.Loading
+        Thread {
+            sleep(5000)
+            liveDataToObserve.postValue(AppState.Success(repoImpl.getWeatherFromLocalStorage()))
+        }.start()
+    }
+}
