+++ app/src/main/java/com/example/weather/view/MainActivity.kt
@@ -1,0 +1,22 @@
+package com.example.weather.view
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.weather.R
+import com.example.weather.databinding.MainActivityBinding
+
+class MainActivity : AppCompatActivity() {
+    private lateinit var binding: MainActivityBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = MainActivityBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        if (savedInstanceState == null) {
+            supportFragmentManager.beginTransaction()
+                    .replace(R.id.container, MainFragment.newInstance())
+                    .commitNow()
+        }
+    }
+}
