+++ app/src/main/res/values/colors.xml
@@ -7,5 +7,5 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="loadingBackground">#00FFFFFF</color>
+    <color name="loading_background">#00FFFFFF</color>
 </resources>
