+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="frame_layout_margin">10dp</dimen>
+    <dimen name="city_name_margin_top">30dp</dimen>
+    <dimen name="city_name_text_size">35sp</dimen>
+    <dimen name="city_coordinates_margin_top">5dp</dimen>
+    <dimen name="default_margin_top">15dp</dimen>
+    <dimen name="temperature_value_text_size">40sp</dimen>
+    <dimen name="feels_like_text_size">30sp</dimen>
+</resources>
