+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,6 @@
 <resources>
     <string name="app_name">Weather</string>
+    <string name="city_coordinates_text">lt/ln: </string>
+    <string name="temperature_label_text">Температура</string>
+    <string name="feels_like_label">Ощущается как</string>
 </resources>
